+++ src/input/type/QuickPick.ts
@@ -5,10 +5,28 @@ import * as vscode from "vscode";
  * The type of the function to generate the items for the quickPick.
  *
  * This function can read all previous entered dialog values.
+ *
+ * @returns an `QuickPickItem` array with all the loaded items. This can be returned as a normal element, or packed with an additional title.
  */
 export type QuickPickItemFunction = (
   currentResults: DialogValues
-) => Promise<vscode.QuickPickItem[]> | vscode.QuickPickItem[];
+) => Promise<QuickPickItems> | Promise<vscode.QuickPickItem[]> | vscode.QuickPickItem[];
+
+/**
+ * The items loaded from the `QuickPickItemFunction`.
+ * You should only use this, if you want to have an additional title.
+ */
+export interface QuickPickItems {
+  /**
+   * The items itself.
+   */
+  items: vscode.QuickPickItem[];
+
+  /**
+   * Any additional title. If this title was given, it will be displayed in brackets in the title of the quick pick.
+   */
+  additionalTitle?: string;
+}
 
 /**
  * Any quick pick that is not an selection of an connection.
@@ -48,13 +66,16 @@ export class QuickPick extends InputBase {
     currentStep: number,
     maximumStep: number
   ): Promise<string[] | undefined> {
-    const items = await this.generateItems(currentResults);
+    const items = await this.loadGeneratedItems(currentResults);
 
-    const result: vscode.QuickPickItem | vscode.QuickPickItem[] | undefined = await vscode.window.showQuickPick(items, {
-      canPickMany: this.allowMultiple,
-      title: this.generateTitle(this.title, currentStep, maximumStep),
-      placeHolder: this.generatePlaceholder(),
-    });
+    const result: vscode.QuickPickItem | vscode.QuickPickItem[] | undefined = await vscode.window.showQuickPick(
+      items.items,
+      {
+        canPickMany: this.allowMultiple,
+        title: this.generateTitle(this.title, currentStep, maximumStep, items.additionalTitle),
+        placeHolder: this.generatePlaceholder(),
+      }
+    );
 
     if (result) {
       if (Array.isArray(result)) {
@@ -68,14 +89,42 @@ export class QuickPick extends InputBase {
   }
 
   /**
+   * Loads the items via the `generateItems` function.
+   * It will also transform all the different data formats into one format.
+   * @param currentResults - the current dialog results
+   * @returns the loaded items and an optional additional title
+   */
+  protected async loadGeneratedItems(currentResults: DialogValues): Promise<QuickPickItems> {
+    const items: vscode.QuickPickItem[] = [];
+    let additionalTitle: string | undefined;
+
+    const generatedItems = await this.generateItems(currentResults);
+
+    if (Array.isArray(generatedItems)) {
+      items.push(...generatedItems);
+    } else {
+      items.push(...generatedItems.items);
+      additionalTitle = generatedItems.additionalTitle;
+    }
+
+    return { items, additionalTitle };
+  }
+
+  /**
    * Generates the whole title for the input by using the given title and the step output
    * @param pTitle - the describing text title. This should be given by creating the class
    * @param currentStep - the current step of the input
    * @param maximumStep - the maximum step of the input
+   * @param pAdditionalTitle  - any additional title
    * @returns the generated title
    */
-  protected generateTitle(pTitle: string, currentStep: number, maximumStep: number): string {
-    return `${pTitle} -  ${this.generateStepOutput(currentStep, maximumStep)}`;
+  protected generateTitle(pTitle: string, currentStep: number, maximumStep: number, pAdditionalTitle?: string): string {
+    let generatedTitle = pTitle;
+    if (pAdditionalTitle) {
+      generatedTitle += ` (${pAdditionalTitle})`;
+    }
+    generatedTitle += ` -  ${this.generateStepOutput(currentStep, maximumStep)}`;
+    return generatedTitle;
   }
 
   /**